requirements.txt
requests
rich
llama-cpp-python
